Main: - Created AddedList Component, - Created FillForm Component, - Created useSearchRepo Hook, - Refactore all Main Page
